requirements.txt
streamlit
phidata
dotenv